pyproject.toml
[tool.poetry]
name = "demo-digitalform-api"
version = "0.1.0"
description = ""
authors = ["Natthasath Saksupanara <natthasath.sak@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "demo_digitalform_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
fastapi-utils = "^0.2.1"
uvicorn = "^0.20.0"
python-decouple = "^3.6"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.10.2"}
fpdf2 = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
